+++ thirdswaicall-20251127/alibabacloud_thirdswaicall20251127/models/_query_task_concurrency_request.py
@@ -8,9 +8,13 @@ class QueryTaskConcurrencyRequest(DaraModel):
     def __init__(
         self,
         application_code: str = None,
+        caller_uac_account_id: str = None,
+        current_workspace_id: str = None,
         task_id: int = None,
     ):
         self.application_code = application_code
+        self.caller_uac_account_id = caller_uac_account_id
+        self.current_workspace_id = current_workspace_id
         self.task_id = task_id
 
     def validate(self):
@@ -24,6 +28,12 @@ class QueryTaskConcurrencyRequest(DaraModel):
         if self.application_code is not None:
             result['ApplicationCode'] = self.application_code
 
+        if self.caller_uac_account_id is not None:
+            result['CallerUacAccountId'] = self.caller_uac_account_id
+
+        if self.current_workspace_id is not None:
+            result['CurrentWorkspaceId'] = self.current_workspace_id
+
         if self.task_id is not None:
             result['TaskId'] = self.task_id
 
@@ -33,6 +43,12 @@ class QueryTaskConcurrencyRequest(DaraModel):
         m = m or dict()
         if m.get('ApplicationCode') is not None:
             self.application_code = m.get('ApplicationCode')
+
+        if m.get('CallerUacAccountId') is not None:
+            self.caller_uac_account_id = m.get('CallerUacAccountId')
+
+        if m.get('CurrentWorkspaceId') is not None:
+            self.current_workspace_id = m.get('CurrentWorkspaceId')
 
         if m.get('TaskId') is not None:
             self.task_id = m.get('TaskId')